+++ messages.go
@@ -1,8 +1,9 @@
 package main
 
 type AcceptMessage struct {
-	Type       string        `json:"type"`
-	IceServers []interface{} `json:"iceServers,omitempty"`
+	Type        string        `json:"type"`
+	IceServers  []interface{} `json:"iceServers,omitempty"`
+	IsExistUser bool          `json:"isExistUser"`
 }
 
 type RejectMessage struct {
@@ -11,7 +12,8 @@ type RejectMessage struct {
 }
 
 type AcceptMetadataMessage struct {
-	Type       string        `json:"type"`
-	Metadata   interface{}   `json:"authzMetadata,omitempty"`
-	IceServers []interface{} `json:"iceServers,omitempty"`
+	Type        string        `json:"type"`
+	Metadata    interface{}   `json:"authzMetadata,omitempty"`
+	IceServers  []interface{} `json:"iceServers,omitempty"`
+	IsExistUser bool          `json:"isExistUser"`
 }